refactor(bulk fastt): move grid editor to different file
